requirements.txt
streamlit
pandas
sqlite3
requests
numpy
Pillow
scikit-learn
httpx
beautifulsoup4
huggingface_hub